Cleaned up some of teams code, moved to API and worked on modals/notifications
