+++ backend/api/base/modal.py
@@ -1,7 +1,7 @@
 from django.http import HttpRequest, HttpResponseBadRequest
 from django.shortcuts import render
 
-from backend.models import UserSettings, Invoice
+from backend.models import UserSettings, Invoice, Team
 
 
 # Still working on
@@ -19,6 +19,11 @@ def open_modal(request: HttpRequest, modal_name, context_type=None, context_valu
                     )
                 except UserSettings.DoesNotExist:
                     pass
+            elif context_type == "accept_invite_with_code":
+                context["code"] = context_value
+            elif context_type == "team_leave":
+                if request.user.teams_joined.filter(id=context_value).exists():
+                    context["team"] = Team.objects.filter(id=context_value).first()
             elif context_type == "edit_invoice_to":
                 invoice = context_value
                 try:
@@ -43,5 +48,6 @@ def open_modal(request: HttpRequest, modal_name, context_type=None, context_valu
                     # context["to_country"] = invoice.client_country
 
         return render(request, template_name, context)
-    except:
+    except Exception as e:
+        print(f"Something went wrong with loading modal {modal_name}. Error: {e}")
         return HttpResponseBadRequest("Something went wrong")
