+++ backend/models.py
@@ -367,6 +367,8 @@ class Notification(models.Model):
     message = models.CharField(max_length=100)
     action = models.CharField(max_length=10, choices=action_choices, default="normal")
     action_value = models.CharField(max_length=100, null=True, blank=True)
+    extra_type = models.CharField(max_length=100, null=True, blank=True)
+    extra_value = models.CharField(max_length=100, null=True, blank=True)
     date = models.DateTimeField(auto_now_add=True)
