+++ frontend/templates/base/_notification_dropdown_items.html
@@ -21,7 +21,7 @@
                hx-trigger="click once"
                hx-swap="beforeend"
                hx-target="#modal_container"
-               hx-get="{% url "api:base:modal retrieve" modal_name=notification.action_value %}">
+               hx-get="{% url "api:base:modal retrieve with context" modal_name=notification.action_value context_type=notification.extra_type context_value=notification.extra_value %}">
             {{ notification.message }}
         </label>
     </li>
